Conflicts:
	app/scripts/app.js
	app/scripts/services/resources-service.js

+++ app/scripts/controllers/resources-controllers.js
@@ -9,19 +9,21 @@
 
     vm.list = {};
 
-
+    // resource operations
     vm.stopResource = stopResource;
     vm.startResource = startResource;
     vm.restartResource = restartResource;
     vm.deleteResource = deleteResource;
+    vm.isOperationAvailable = isOperationAvailable;
 
+    // search
     vm.searchInput = '';
     vm.search = search;
+
+    // pagination
     vm.changePageSize = changePageSize;
     vm.changePage = changePage;
     vm.getNumber = getNumber;
-
-
     vm.pageSizes = [1,5,10,15,20];
     vm.currentPageSize = resourcesService.pageSize;
     vm.pages = resourcesService.pages ? resourcesService.pages : 5;
@@ -30,7 +32,7 @@
     getResourceList();
 
     function getResourceList() {
-      resourcesService.getResourcesList().then(function (response) {
+      resourcesService.getResourcesList().then(function(response) {
         vm.pages = resourcesService.pages;
         vm.list = response;
       });
@@ -56,9 +58,14 @@
 
 
     function deleteResource(uuid, index) {
-      resourcesService.deleteResource(uuid).then(function(response){
-        vm.list.splice(index, 1);
-      });
+      var confirmDelete = confirm('Confirm resource deletion?');
+      if (confirmDelete) {
+        resourcesService.deleteResource(uuid).then(function(response){
+          vm.list.splice(index, 1);
+        });
+      } else {
+        alert('Resource was not deleted.');
+      }
     }
 
     function search() {
@@ -84,6 +91,13 @@
     function getNumber(num) {
       return new Array(num);
     }
+
+    function isOperationAvailable(resource, operation) {
+      var availableOperations = resourcesService.getAvailableOperations(resource);
+      operation = operation.toLowerCase();
+      return availableOperations.indexOf(operation) !== -1;
+    }
+
   }
 
 })();
@@ -91,23 +105,20 @@
 (function() {
   angular.module('ncsaas')
     .controller('ResourceAddController',
-      ['$state', 'resourcesService', 'cloudsService', 'projectsService', 'keysService', 'templatesService',
+      ['$state', 'resourcesService', 'servicesService', 'projectsService', 'keysService', 'templatesService',
       ResourceAddController]);
 
   function ResourceAddController(
-      $state, resourcesService, cloudsService, projectsService, keysService, templatesService) {
+      $state, resourcesService, servicesService, projectsService, keysService, templatesService) {
     var vm = this;
 
     // Resource add process:
     // 0. User has to choose back-end. Currently this step is not active because we have only one back-end.
-    // 1. User has to choose project.
+    // 1. User has to choose project. Method setProject has to be called, because services are initiated based on
+    // projects
     // 2. User has to choose service. Method setService has to be called, because flavors and templates are initialized
     // based on services
     // 3. User has to choose flavor, template, ssh key and other resource attributes.
-
-    vm.selectedService;
-    vm.selectedTemaplate;
-    vm.selectedFlavor;
 
     vm.showFlavors = false;
     vm.showTemplates = false;
@@ -123,6 +134,7 @@
     vm.setProject = setProject;
     vm.resource = resourcesService.createResource();
     vm.save = save;
+    vm.cancel = cancel;
     vm.errors = {};
 
     function activate() {
@@ -159,7 +171,7 @@
     function setProject(project) {
       vm.resource.project = project.url;
       // services
-      cloudsService.getCloudList({project: project.uuid}).then(function(response) {
+      servicesService.getServiceList({project: project.uuid}).then(function(response) {
         vm.serviceList = response;
       });
     }
@@ -174,6 +186,10 @@
       function error(response) {
         vm.errors = response.data;
       }
+    }
+
+    function cancel() {
+      $state.go('resources');
     }
 
     activate();
